bit of a refactor and it closes #7
it also adds the days until codeday and a spotify card on the day-of as well as some John Peter announcements and it blocks the check-in tab until codeday

+++ CodeDay Companion/API/Models/Event.swift
@@ -11,8 +11,8 @@ struct Event : Codable {
     var region: String
     var name: String
     var schedule: Dictionary<String, [Activity]>
-    var startsAt: Int
-    var endsAt: Int
+    var startsAt: Double
+    var endsAt: Double
     var venue: Venue?
     
     enum CodingKeys : String, CodingKey {
